refactor js to use python templates
